Create Analysis
Analysis with the use of python. During the analysis, I imported  libraries (seaborn, matplotlib), inserted a csv file with a database of a company's sales and expenditure for the year 2018.
The code reads the data from the spreadsheet, collects all the sales from each month into a single list and used bar charts to represent the data for sales and expenditure for each month of year 2018.